[#1180] remove finalizer usage and move app reconcile logic into broker service controller as it owns the app secret that is the source of truth for the service config
