+++ bundle/manifests/arkmq-org-broker-operator.clusterserviceversion.yaml
@@ -5933,7 +5933,6 @@ spec:
           - activemqartemises
           - activemqartemisscaledowns
           - activemqartemissecurities
-          - brokerapps
           - brokerservices
           verbs:
           - create
@@ -5950,7 +5949,6 @@ spec:
           - activemqartemises/finalizers
           - activemqartemisscaledowns/finalizers
           - activemqartemissecurities/finalizers
-          - brokerapps/finalizers
           - brokerservices/finalizers
           verbs:
           - update
@@ -5961,7 +5959,6 @@ spec:
           - activemqartemises/status
           - activemqartemisscaledowns/status
           - activemqartemissecurities/status
-          - brokerapps/status
           - brokerservices/status
           verbs:
           - get
