+++ config/rbac/role.yaml
@@ -57,7 +57,6 @@ rules:
   - activemqartemises
   - activemqartemisscaledowns
   - activemqartemissecurities
-  - brokerapps
   - brokerservices
   verbs:
   - create
@@ -74,7 +73,6 @@ rules:
   - activemqartemises/finalizers
   - activemqartemisscaledowns/finalizers
   - activemqartemissecurities/finalizers
-  - brokerapps/finalizers
   - brokerservices/finalizers
   verbs:
   - update
@@ -85,7 +83,6 @@ rules:
   - activemqartemises/status
   - activemqartemisscaledowns/status
   - activemqartemissecurities/status
-  - brokerapps/status
   - brokerservices/status
   verbs:
   - get
