+++ deploy/activemq-artemis-operator.yaml
@@ -9597,7 +9597,6 @@ rules:
   - activemqartemises
   - activemqartemisscaledowns
   - activemqartemissecurities
-  - brokerapps
   - brokerservices
   verbs:
   - create
@@ -9614,7 +9613,6 @@ rules:
   - activemqartemises/finalizers
   - activemqartemisscaledowns/finalizers
   - activemqartemissecurities/finalizers
-  - brokerapps/finalizers
   - brokerservices/finalizers
   verbs:
   - update
@@ -9625,7 +9623,6 @@ rules:
   - activemqartemises/status
   - activemqartemisscaledowns/status
   - activemqartemissecurities/status
-  - brokerapps/status
   - brokerservices/status
   verbs:
   - get
