+++ deploy/cluster_role.yaml
@@ -55,7 +55,6 @@ rules:
   - activemqartemises
   - activemqartemisscaledowns
   - activemqartemissecurities
-  - brokerapps
   - brokerservices
   verbs:
   - create
@@ -72,7 +71,6 @@ rules:
   - activemqartemises/finalizers
   - activemqartemisscaledowns/finalizers
   - activemqartemissecurities/finalizers
-  - brokerapps/finalizers
   - brokerservices/finalizers
   verbs:
   - update
@@ -83,7 +81,6 @@ rules:
   - activemqartemises/status
   - activemqartemisscaledowns/status
   - activemqartemissecurities/status
-  - brokerapps/status
   - brokerservices/status
   verbs:
   - get
